+++ babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  presets: ['@babel/preset-typescript', '@babel/preset-env'],
+  ignore: ['src/__tests__/*.test.ts', 'src/__testfixtures__/**/*.ts', 'plop-templates/*'],
+};