+++ src/__testfixtures__/convert-template-literal.input.js
@@ -1,0 +1,3 @@
+const a = `abcde`;
+const b = `1234${true}`;
+const c = `abc 123`;