+++ src/__testfixtures__/remove-debugger.input.js
@@ -1,0 +1,8 @@
+1 + 1;
+debugger;
+1 + 1;
+
+{
+  debugger;
+  1 + 1;
+}