+++ CHANGELOG.md
@@ -1,3 +1,84 @@
+# [1.10.0](https://github.com/ticklepoke/codemods/compare/v1.9.0...v1.10.0) (2021-06-12)
+
+
+### Features
+
+* **transform:** Add template literals ([a939d9f](https://github.com/ticklepoke/codemods/commit/a939d9fa9f076c1556e186ceb89f03ac20f36b58))
+
+# [1.9.0](https://github.com/ticklepoke/codemods/compare/v1.8.0...v1.9.0) (2021-06-12)
+
+
+### Features
+
+* **terminal colors:** Add terminal colors ([4dbb3d4](https://github.com/ticklepoke/codemods/commit/4dbb3d442bb2bc0f85ebf9abb886895361077516))
+
+# [1.8.0](https://github.com/ticklepoke/codemods/compare/v1.7.0...v1.8.0) (2021-06-12)
+
+
+### Features
+
+* **babel:** Add babel support ([eeb6c9b](https://github.com/ticklepoke/codemods/commit/eeb6c9ba991ccdae2acfae181e1a82082df0d429))
+* **dist:** Add dist files ([e9a50d6](https://github.com/ticklepoke/codemods/commit/e9a50d6ca76550b42be7e8e28315e8b3ace6e589))
+* **plop:** Add plopfile command ([2fccffd](https://github.com/ticklepoke/codemods/commit/2fccffde42a84c22585191c87e58829019c13545))
+
+# [1.7.0](https://github.com/ticklepoke/codemods/compare/v1.6.0...v1.7.0) (2021-06-11)
+
+
+### Features
+
+* **transform:** Add remove-console transform ([11a9edf](https://github.com/ticklepoke/codemods/commit/11a9edf9cdf6d8c94dae2653d3b8470cfafa6180))
+* **utils:** Add multiple transform util ([a1a1e51](https://github.com/ticklepoke/codemods/commit/a1a1e516032f547a8263c1193541e80279932b5b))
+
+# [1.6.0](https://github.com/ticklepoke/codemods/compare/v1.5.0...v1.6.0) (2021-06-11)
+
+
+### Features
+
+* **transform:** Add remove-debugger transform ([99f21bf](https://github.com/ticklepoke/codemods/commit/99f21bf5d4c81e5e2576fc6e3f7440cc57fdb058))
+
+# [1.5.0](https://github.com/ticklepoke/codemods/compare/v1.4.0...v1.5.0) (2021-06-11)
+
+
+### Features
+
+* **plop:** Add transform plop ability ([311dcc0](https://github.com/ticklepoke/codemods/commit/311dcc0bb2099cdde8f69698d2d8e778be4e0cdd))
+
+# [1.4.0](https://github.com/ticklepoke/codemods/compare/v1.3.0...v1.4.0) (2021-06-11)
+
+
+### Bug Fixes
+
+* **jest:** Ignore plop templates ([d7637a3](https://github.com/ticklepoke/codemods/commit/d7637a301ca87b5b4acb6e8bf848fa145b22fc0f))
+
+
+### Features
+
+* **plop:** Add plop file ability ([ec02999](https://github.com/ticklepoke/codemods/commit/ec029998d6c386ff6a5e8b6954f7a8c11d4a298b))
+
+# [1.3.0](https://github.com/ticklepoke/codemods/compare/v1.2.0...v1.3.0) (2021-06-11)
+
+
+### Bug Fixes
+
+* **eslint:** Single quotes ([490a5e3](https://github.com/ticklepoke/codemods/commit/490a5e34eafa9ee0fe9ba8ba22294e710ccedcdf))
+
+
+### Features
+
+* **husky:** Setup husky, lintstaged ([ff620e0](https://github.com/ticklepoke/codemods/commit/ff620e047b9bf0bedafd40d4f6cf25b149f8919a))
+
+# [1.2.0](https://github.com/ticklepoke/codemods/compare/v1.1.1...v1.2.0) (2021-06-11)
+
+
+### Bug Fixes
+
+* **ci:** Fix github credentials ([8a02b8b](https://github.com/ticklepoke/codemods/commit/8a02b8bdc73659039dc518858235902ab81eedc6))
+
+
+### Features
+
+* **jscodeshift:** Adds jscodeshift ([a82dbda](https://github.com/ticklepoke/codemods/commit/a82dbda20413001a5ba85f1f92288f6acc17f050))
+
 ## [1.1.1](https://github.com/ticklepoke/codemods/compare/v1.1.0...v1.1.1) (2021-06-11)